Redefine main menu structure to get rid of sub menus.
