+++ gamestonk_terminal/options/tradier_model.py
