+++ gamestonk_terminal/options/yfinance_model.py
